I created insertExistingMessage
insertExistingMessage(in ARVC+UserGesture) should take location and altitude (in the form of raycastResult)of a message a user created in insertNewMessage. The reason we use raycastResult is because it contains the location and altitude. The reason we keep it in raycastResult format instead of just using location and altitude as separate parameters is because raycastResult has built-in features like raycastResult.worldTransform in line 166 of ARVC+UserGestures