Added ability to place markers/messages in place (#52)
* Built test functions for mock data

* Working on saving and loading AR messages

I am currently saving the user tap locations in an array so I can populate existing messages based on location of user taps.

* Got data points of AR message

Progress: I placed an AR message outside my house to get the location and altitude. Next I want to use this to simulate loading an existing message to our AR world from servers.

Setback: In line 180 in CGRect() there is a parameter called touchLocation which tells our AR world where to place message. Since a user is not clicking it due to use loading the message from our server, what should I replace "touchLocation" with?

* I created insertExistingMessage

insertExistingMessage(in ARVC+UserGesture) should take location and altitude (in the form of raycastResult)of a message a user created in insertNewMessage. The reason we use raycastResult is because it contains the location and altitude. The reason we keep it in raycastResult format instead of just using location and altitude as separate parameters is because raycastResult has built-in features like raycastResult.worldTransform in line 166 of ARVC+UserGestures

* In Progess: Save and load existing messages

* Adding Menu Button for Testing (In-Progress)

I am adding a menu to our ARView in mainstoryboard. The reason I am adding it is to test the save and load geoAnchors feature from server. For now I am goin to save it locally on the phone, later on connect to server.

* Added Save/Load button to ARView for testing

Save/Load button will ask the user if they want to save or load anchors in our ARView. I want to test and make sure I can save and load anchors, to and from server.

* In Progress: Working on Loading/Saving GeoAnchors

* In Progress: ARGeoAnchor existing

* Complete: Placing and Loading AR Messages Works

To test code feature I made a temporary Load/Save button on the AR page. You post messages -> Save Anchors -> Delete all the anchors -> then Load all the anchors

In progess: We need to further test why the AR messages load to random spots.

* Pushing code before tweaking Inserting Messages

* filp flip flip

* dev: added ability to place markers in place

Co-authored-by: CarlosLoeza <cloeza1@mail.ccsf.edu>